Make the Carpet WaveToy almost-work with two refinement levels.  The
Make the Carpet WaveToy almost-work with two refinement levels.  The
basic error was that the Carpet time levels were set to the Cactus
time levels, but they need to be different.  In Carpet, level 0 is
always the current level, and not so in Cactus.

darcs-hash:20010312155417-f6438-7c09703fc594525f2d68d4f9da2e3f3eeaec9c57.gz