This is a little Test thorn for progressive meshrefinement. It sets up a Gaussian scalar field and
This is a little Test thorn for progressive meshrefinement. It sets up a Gaussian scalar field and
measures the difference between the data on the finite difference grids and the analytical values. It
still needs some work, but the basic stuff workes. Right now it is setup for ENO prolongation. If you
want to change the way prolongation is handled, you have to edit 'param.ccl'

- Christian

----------------------------------------------------------------------

darcs-hash:20040523135614-19929-f2251e59534aae6da120a0d95135d980bbd4a4f4.gz