Add a new tags table entry to check: InterpNumTimelevels.
Add a new tags table entry to check: InterpNumTimelevels.
When this is set then the interpolator will take the given number of
timelevels as the number to use when interpolating in time. So if you
have grid functions with fewer than 3 timelevels mixed with grid
functions with 3 timelevels you can use this tag, set it to the number
of timelevels that you have available, and the interpolator will "do
the right thing".

Note that this means that GFs with only one timelevel are implicitly
treated as being constant in time.

The obvious examples, soon to be committed (I hope) are the static
conformal factor and the coordinates (although why you'd want to
interpolate the coordinates...). E.g.,

REAL confac TYPE = GF timelevels = 1 tags='tensortypealias="Scalar"
Prolongation="None" InterpNumTimelevels=1'

...

darcs-hash:20040820072138-58737-5ff68c487af14a1d72228aa6697cc9eea1ae3a89.gz