CarpetLib: Collect more timing statistics
Collect more timing statistics in the data class.  Print these
statistics to stdout when the Cactus parameter print_timestats is set.

Create a timer class "timestat".  This is a timer that can be started,
stopped, and it prints the total time as well as some statistics.

For memory allocation statistics, count the number of objects as well
as the number of bytes.

darcs-hash:20041230212136-891bb-c14edfa7d539ae9b135eee76afadaad51fd0b098.gz