global: Turn CarpetLib templates into classes
Turn most of the templates in CarpetLib, which used to have the form
        template<int D> class XXX
into classes, i.e., into something like
        class XXX
by setting D to the new global integer constant dim, which in turn is set to 3.

The templates gf and data, which used to be of the form
        template<typename T, int D> class XXX
are now of the form
        template<typename T> class XXX

The templates vect, bbox, and bboxset remain templates.

This change simplifies the code somewhat.

darcs-hash:20050101182234-891bb-c3063528841f0d078b12cc506309ea27d8ce730d.gz