global: Change the way in which the grid hierarchy is stored
Change the way in which the grid hierarchy is stored.  The new hierarchy is
        map
        mglevel
        reflevel
        component
        timelevel
i.e., mglevel moved from the bottom to almost the top.  This is
because mglevel used to be a true multigrid level, but is now meant to
be a convergence level.

Do not allocate all storage all the time.  Allow storage to be
switched on an off per refinement level (and for a single mglevel,
which prompted the change above).  Handle storage management with
CCTK_{In,De}creaseGroupStorage instead of
CCTK_{En,Dis}ableGroupStorage.

darcs-hash:20050201225827-891bb-eae3b6bd092ae8d6b5e49be84c6f09f0e882933e.gz