Adaptive regridding symmetry / indexing partial fix
Fixed a bug where the array indexing in C++ was wrong. Now something
that should be symmetric almost is.

There is still an assertion failure when it tries to derefine - for
the moment this appears to be a Carpet internal problem.

darcs-hash:20050209161923-58c7f-7d13dd7c72c44d935bff9050e3d9281f5f39ec90.gz