CarpetLib: Do not check invariant of bboxset class
Checking the invariant of the bboxset class is probably O(N^3) in the
number of bboxes, possibly worse.  It is a very slow operation when
there are many components in a simulation, especially with AMR, and
possibly also when running on many processors.

darcs-hash:20050321020814-891bb-c5eb1cde6f3ac064e39a362b19bd20e15d03bc24.gz