CarpetRegrid: Better error messages for wrong refinement specifications
Check that there are no more than Carpet::max_refinement_levels-1
refinement levels specified, and abort with an error message if so.

When aligning the grid size to the outer boundary, align lower and
upper boundary at the same time.  Otherwise there is an intermediate
grid size which may not be a multiple of the grid spacing.

darcs-hash:20050601144240-891bb-a558598231e19322eba7f02339b767c92c603edf.gz