CarpetReduce: Handle empty domains gracefully
Empty simulation domains occur e.g. with Cartoon2D, where the total
volume of the domain is zero: All evolved grid points lie on the
boundary.  Before, CarpetReduce did not handle this special case
correctly.

Instead of returning 0 for all reductions, set up the grid point
weights to perform a 2D reduction on the data.  That is, the domain is
treated as if it was one grid point wide in the direction in which it
has the extent zero.

darcs-hash:20050603220813-891bb-83088cbd570d967aee95170915c54d905ea5bf71.gz