CarpetInterp: break routine into pieces
This is the first stage of a breakup of the monster function
Carpet_DriverInterpolate

The routine is segmented into several functional subroutines, with
names given (mostly) by the one-line comments that were in the
original.

Todo:

1) One of these new subroutines, DoLocalInterpolation, is itself a
monster function, and needs to be dealt with.

2) All of the subroutines have very long argument lists.  There are
repeated patterns of arguments, which suggest that a structure could
be created which embodies the pattern.

darcs-hash:20050715103538-891bb-c0757adc59d29b09c63b7e4dbe11b57cd46b6682.gz