CarpetLib: Disable check for time extrapolation for op_copy grid functions
Disable the check for time extrapolation for grid functions that are
not interpolated but copied (transport_operator=op_copy).  This check
needs to know the sign of delta_time, and that sign is not known here.
(For example, 3-timelevel-initialisation evolves both forwards and
backwards in time.)

darcs-hash:20050726111122-891bb-0c2130f2dfdd86c25575cbb88cb7c8ad80089eac.gz