CarpetLib: give data an << operator
Added a ostream & << operator for the data class.
Made an explicit instantiation for CCTK_REAL.

Usually the amount of data renders such an operator useless,
but in test programs it is handy to be able to just print the data values.

darcs-hash:20050809093318-90671-dd30c55dc8757efa96a9bfbe90a48ff445e26f07.gz