CarpetLib: Correct the way prolongation boundaries are determined
Correc the way in which it is determined which points are boundary
prolongated and which are synchronised.

For this, introduce a new field "owned" to the dh class that
determines which grid points a processor owns for interprocessor
communication.

darcs-hash:20061113214128-dae7b-34a520f7e461173d0042acde68cb48cd670b167a.gz