CarpetRegrid2: Ensure and test proper nesting
Introduce a parameter ensure_proper_nesting, which automatically
enlarges the levels of the grid hierarchy to ensure proper nesting.
(It does not enlarge the coarsest level.)  If this parameter is not
set, proper nesting is still tested, and a warning is emitted if it is
violated.

darcs-hash:20070425002649-dae7b-7431881af786bbae5048c5329807e90a4f7b2e29.gz