CarpetReduce: Implement pointwise reduction operators
Implement a new set of reduction operators for Carpet.  These have the
same definitions as the old operators, except that they give the same
weight to all grid points.  The existing operators give a larger
weight to coarser grid points, so that physical reduction operators
work correctly, e.g. for calculating the total mass.

The new operators have the same name as the old operators plus a
prefix "i" (the lower case letter "eye").  For example, the new L2
norm is called "inorm2", while the old norm continues to be called
"norm2".

The physical (old) operators converge to a result which is independent
of the grid structure.  The pointwise (new) operators still converge,
but to a value which depends on the grid structure.

In order to make it possible to compare situations with and without
symmetries, boundary points still carry a weight of 1/2 or whatever is
appropriate.  That means that there exists a simple relationship
between the pointwise norms with and without a symmetry boundary.

darcs-hash:20070526195016-dae7b-6a7e2a39bebd9c0a6377f908a4d7f446f8288191.gz