Carpet: Call global mode initialisation routines on finest level
When setting up initial data or when recovering, execute global mode
routines not with the coarsest level, but either for every level or
for the finest level.  Execute them on every level if the refinement
hierarchy can change while iterating over the levels.  If the
hierarchy cannot change, execute them with the finest level, which is
the last level which is treated.

This order is more similar to the way in which global mode routines
are handled during evolution.  It ensures e.g. that
CarpetReduce::weight is set up correctly.

darcs-hash:20070531215139-dae7b-ef900bee28ab52dd410366dc22f26d37dba37a86.gz