Add capability to use Fletcher32 checksums for HDF5 files.
The idea is that data corruption while writing, storing and reading the
data should be detected by this mechanism, preventing restarts from
corrupt checkpoints for example.

To activate this, set the new parameter "IOHDF5::use_checksums" to
"true". The default (for the moment) is false.

Note that the use of compression already gives some kind of checksum
mechanism.