Add a new prolongation operator Lagrange_monotone.
This is meant to be a copy of the prolongation method used in SACRA
(as I understand it) and uses 5th order Lagrangian where the result is
monotone, reducing to first order (linear) where necessary.

Briefly tested. Intended for use with hydro codes - should allow for
mesh boundaries inside the star.