LoopControl: Implement cache-collaborative multi-threading Ignore-this: 5169757c7749834ae595d4d73b39220
Add a new, additional feature to LoopControl: different threads can
work on small regions that are likely to use the same cache entries as
other threads, trying to reduce cache pressure.  This makes sense
mostly when the regions are still expensive although they are small,
e.g. for the BSSN RHS.