Carpet: Implement padding from grid variables Ignore-this: ac4962293cd7c0313a0f17f3ccd59e96
Set the cGH and cGroupDynamicData entries correctly if padding is
used.

cctk_lsh then contains the number of grid points that have been
allocated, cctk_lssh contains the number of grid points that should be
used.  cctk_lssh is also used for staggered grid functions, although
this kind of staggering is not supported by Carpet.