CarpetIOHDF5: Index file support
Scanning the attributes of a large CarpetIOHDF5 output file, as is
necessary in the visitCarpetHDF5 plugin, can be very time consuming.
This commit adds support for writing an "index" HDF5 file at the same
time as the data file, conditional on a parameter
"CarpetIOHDF5::output_index".  The index file is the same as the data
file except it contains null datasets, and hence is very small.  The
attributes can be read from this index file instead of the data file,
greatly increasing performance.  The datasets will have size 1 in the
index file, so an additional attribute (h5space) is added to the
dataset to specify the correct dataset dimensions.