LoopControl: change storage size calculation to work with Intel 11.1
it seems that ifort's INQUIRE statement returns the byte size of a record
divided by four. Maybe it is counting ints? This one uses pointer arithmetic
and will work unless Fortran pads the structures differently than C or if a
compiler allocates some temporaries when members of arrays are passed to
functions.