Make Carpet timers hierarchical
This commit modifies the timer infrastructure in Carpet to generate a
tree of timers where the hierarchy reflects the call-graph of the
program.  This makes it much easier to interpret the timer output than
with the previous flat structure, where it was not possible to see
which timers "contained" which others.  More implementation details
are given at the top of TimerNode.hh.

Note that the Timer source and header files have been renamed as
CactusTimer and a new Timer file and object has been created.  This is
because the Timer object now only provides a wrapper around the Cactus
timer mechanism which was contained in the old Timer object.