CarpetRegrid2: Enforce certain grid hierarchy properties
Enforce certain grid structure properites, and ensure that they hold
after being enforced.

This change may lead to Carpet using a (slightly) different grid
structure than before. The differences are not "important"; what
potentially changes is the order in which consistency conditions are
applied. I expect that, in most cases, there will be no differences.
The list of consistency conditions is:

- ensure proper nesting (level must be larger than next finer level)
- add buffer zones (only once, not really a consistency condition)
- combine regions into a single box (if this is efficient)
- apply rotating 90/180 symmetry (if requested)
- clip at outer boundary (so that boxes that are too large or are
  outside are cut off)

This change also reorganises the code in CarpetRegrid2.