Carpet: Add backtrace output to file on fatal signals Modified version of Ian Hinder's and Justin Luitjens' code.
When a Cactus simulation aborts with a signal, it is often difficult to
determine which part of the code led to the problem. The attached patch
registers a signal handler on Carpet startup for signals 11 and 6
(segmentation fault and abort, e.g. from assert()) which outputs a stack
backtrace from each process to a file, including demangling symbol names. It
uses some low-level and possibly unofficial APIs, and is likely not completely
portable. However, I have tested it on Mac OS (gcc) and Linux (intel) and it
works in those places.

Part of this code was contributed by Justin Luitjens at the Carpet developers'
workshop in summer 2010.