Add Doxygen configuration files
Doxygen is a program which generates HTML documentation describing all
the files, functions and classes of a program based on the C++ source.
It uses comments in a particular format in the code. There is a search
box and a tree view for the classes, as well as links to HTML versions
of the source files. An example of the output is at
  http://damiana2.aei.mpg.de/~ianhin/carpet-doc/

To generate the documentation, go into the root Carpet directory and
type "doxygen", assuming that doxygen is installed on the local
system. The documentation will be generated in a new carpet-doc
directory. Open the file carpet-doc/html/index.html to browse the
documentation.

It would be convenient to have the documentation generated in this way
made available on the Carpet web site and updated after each commit to
the repository. This patch provides a doxygen configuration file and a
header file for the main documentation page.