Fix Carpet2ygraph.pl to work with new Carpet
The script CarpetIOASCII/src/util/Carpet2ygraph.pl (it is the same
both in the git and hg versions of Carpet) is not working anymore. I
modified it in order to be able again to produce different output for
the different refinement levels and now it can also work with more
than 9 refinement levels. I have attached my version and the diff with
the version in the repository.