Introduce dummy Carpet::poison_value parameter
This parameter was removed in the past and is not used in current
Carpet.  However, to stop old parameter files from breaking, we
reintroduce it.  There is a warning in the param.ccl file that the
parameter is unused, but no warning at runtime.  This is not a big
problem since poison is only used for debugging.