CarpetIOHDF5: additional check ensuring that a default request is used if the current request does not exist for one_file_per_group output
this is re #410:
--8<-- by Erik Schnetter --8<--
The corresponding code in CarpetIOHDF5.cc, which outputs data that are not
slices, uses the same algorithm. However, it contains an additional check
ensuring that a default request is used if the corresponding request does not
exist. Look for calls to IOUtil_DefaultIORequest to find this code. I believe
that an equivalent logic would correct this problem in OutputSlice.cc.
--8<-- by Erik Schnetter --8<--