Add parameter sync_barriers to insert an MPI barrier before every sync
This causes each process to wait for all the others to reach this point in the execution.  This will happen anyway (for most of the processes) during the sync, and adding this barrier means that this time is not included in the sync timer.  Also add a new timer to time this barrier.