CarpetIOHDF5: make index files sparse files with full datasets
rather than datasets of extent [1,1,1]. It so happens that this patch removes
the separate shape arrays and separate dataspaces for the datasets in index
files and creates the index file datasets with the same dataspace as the
"heavy" file datasets. It retains the h5shape attribute that was originally
introduced for index files even though it is now redundant (since one could
call H5Sget_simple_extent_dims on the datasets in the index files).