CarpetIOHDF5: use simple_extents rather than storage size on disk
storage size is literally the amount of space used on disk, so if eg
compression is used, this is much smaller than the amount of space
required to hold the data in memory.

Also change type of data read in to what the memory dataspace is not
what the dataspace on disk is. This way HDF5 actually converts from the
on-disk representation to the in-memory one.