CarpetInterp: Don't try to set non-existent WaveMoL::num_timelevels in test parameter files
I can't find any reference to this parameter in the history of WaveMoL since it was imported from CVS, and the history before this point seems to have been lost. Many of the WaveMoL variables have 3 timelevels already, and the test failure that now occurs doesn't look like it would be helped by increasing the existing 1-timelevel variables to 3.