Include string null terminator in HDF5 attributes
Up to now, the "name" attribute was stored as a string, excluding the null
terminator. This meant that when the attributed was read back, it could not
be treated as a standard C string since it was missing the null terminator.

Since strlen returns the length of the string excluding the null terminator,
we must add one additional space for the null-terminator character when
creating the attribute.