CarpetIOHDF5: regenerate test data for CarpetWaveToyRecover_test_4proc
this test is not normally run since it requires 4 processes and we only
run with 1 and 2.

The test is affected by a change of grid structure due to 1258f552.

I have verified that the test passes with the old data if I revert
1258f552. I have verified that -- at least for iteration 64 -- the
differences are just moving of data around in the output files.