CarpetIOHDF5: fix check for old string datatype
The current code contains a check for an old way the grid structure
string was saved: as H5T_NATIVE_CHAR. This check test whether the type
found in a (checkpoint) file is of that type. However, this fails if
H5T_NATIVE_CHAR was different on the machine writing the checkpoint and
the one restoring from it (e.g., big vs. little endian int8).

Instead of checking if the saved type is of type H5T_NATIVE_CHAR, this
patch checks for the new datatype being of class H5T_STRING.

This change makes the two restore testsuites pass on big endian machines
(the checkpoint file in the testsuite was written using the old
mechanism, on a little endian machine).