CarpetLib: make gdata transfer functions static functions
they are used to transfer data whenever a rank is involved in
communication even if it is not the target thus does not actually hold a
copy of the gdata object that the transfer goes to. This now avoids
situations where we have to either create a temporary object or
(illegally) use a NULL this pointer.