CarpetIOHDF5: remove now redundant checkpoint recovery test
this test differed from CarpetWaveToyRecover_test_1proc.par only in that
it used a new grid structure string format. Regenerating that checkpoint
file due to data changes also introduced the new grid structure string
to it. To get the old one one, would have to redo this with a version of
Carpet old enough to generate the old string.